+++ phillydb/testing_utils.py
@@ -1,21 +1,31 @@
-
-def maybe_monkeypatch_response(monkeypatch, pytestconfig, columns=None):
+def maybe_monkeypatch_response(
+    monkeypatch, pytestconfig, carto_rows=None, response_override=None
+):
     if not pytestconfig.getoption("make_request"):
 
         def _fake_get(*args, **kwargs):
-            if "phl.carto.com" in args[0]:
-                return MockCartoResponse(columns=columns)
-            else:
+            if response_override is not None:
+                return MockResponse(response_override)
+            elif "phl.carto.com" in args[0]:
+                return MockCartoResponse(data=carto_rows)
+            elif "https://api.phila.gov/tips/account/" in args[0]:
                 return MockRealEstateTaxRevenueResponse()
+            else:
+                return MockResponse()
 
         monkeypatch.setattr("requests.get", _fake_get)
 
 
-class MockRealEstateTaxRevenueResponse:
-    def __init__(self):
-        self.status_code = 200
-        pass
+class MockResponse:
+    def __init__(self, data=None, status_code=200):
+        self.data = data
+        self.status_code = status_code
 
+    def json(self):
+        return self.data
+
+
+class MockRealEstateTaxRevenueResponse(MockResponse):
     def json(self):
         return {
             "data": {
@@ -27,17 +37,12 @@ class MockRealEstateTaxRevenueResponse:
         }
 
 
-class MockCartoResponse:
-    def __init__(self, columns=None):
-        self.columns = columns if columns else []
-        self.status_code = 200
-        pass
-
+class MockCartoResponse(MockResponse):
     def json(self):
-        columns_dict = {c: "2020-01-01 12:00:00" for c in self.columns}
-        if 'parcel_number' in self.columns:
-            columns_dict['parcel_number'] = '1234'
-        elif 'opa_account_num' in self.columns:
-            columns_dict['opa_account_num'] = '1234'
-        return {"rows": [columns_dict], "fields": columns_dict}
-
+        data = self.data if self.data is not None else [{'opa_account_num': '123'}]
+        if self.data is None:
+            data = [{'opa_account_num': '123'}]
+        elif self.data:
+            return {"rows": data, "fields": data[0]}
+        else:
+            return {'rows': [], 'fields': {}}