+++ tests/test_carto_tables.py
@@ -2,17 +2,7 @@ import pytest
 import requests
 
 from phillydb import carto_tables
-from phillydb.testing_utils import MockCartoResponse
-
-
-def maybe_monkeypatch_response(monkeypatch, pytestconfig, columns=None):
-    columns = columns if columns else []
-    if not pytestconfig.getoption("make_request"):
-
-        def _fake_get(*args, **kwargs):
-            return MockCartoResponse(columns)
-
-        monkeypatch.setattr("requests.get", _fake_get)
+from phillydb.testing_utils import MockCartoResponse, maybe_monkeypatch_response
 
 
 @pytest.fixture(params=carto_tables.__all__)
@@ -22,18 +12,30 @@ def table_obj(request):
     return TableClass()
 
 
+def test_all_tables_list(table_obj, pytestconfig, monkeypatch):
+    output_rows = [{"ABC": "DEF"}]
+    maybe_monkeypatch_response(
+        monkeypatch, pytestconfig, output_rows,
+    )
+    df = table_obj.list(limit=1)
+    if df.empty:
+        raise AssertionError(f"{table_obj.title} failed to return a dataframe.")
+
+
 def test_all_tables_query_by_opa_account_numbers(
     opa_account_numbers, table_obj, pytestconfig, monkeypatch
 ):
-    maybe_monkeypatch_response(
-        monkeypatch,
-        pytestconfig,
+    columns = (
         table_obj.default_columns
-        + table_obj.default_opa_properties_public_joined_columns,
+        + table_obj.default_opa_properties_public_joined_columns
+    )
+    output_rows = [{c: "2020-01-01 12:00:00" for c in columns}]
+    maybe_monkeypatch_response(
+        monkeypatch, pytestconfig, output_rows,
     )
     df = table_obj.query_by_opa_account_numbers(opa_account_numbers=opa_account_numbers)
     if df.empty:
-        raise AssertionError(f"{table_class.title} failed to return a dataframe.")
+        raise AssertionError(f"{table_obj.title} failed to return a dataframe.")
 
 
 def test_all_tables_metadata_urls(
@@ -47,10 +49,11 @@ def test_all_tables_metadata_urls(
 def test_query_by_single_str_column(monkeypatch, pytestconfig):
     property_obj = carto_tables.Properties()
     result_columns = ["location", "parcel_number"]
+    columns = result_columns + property_obj.city_owned_prop_filter_cols
+    output_rows = [{c: "2020-01-01 12:00:00" for c in columns}]
+    output_rows[0]["parcel_number"] = "1234"
     maybe_monkeypatch_response(
-        monkeypatch,
-        pytestconfig,
-        result_columns + property_obj.city_owned_prop_filter_cols,
+        monkeypatch, pytestconfig, output_rows,
     )
     df = property_obj.query_by_single_str_column(
         search_column="location",
@@ -63,11 +66,12 @@ def test_query_by_single_str_column(monkeypatch, pytestconfig):
 
     license_obj = carto_tables.Licenses()
     result_columns = ["licensetype", "opa_account_num"]
-    maybe_monkeypatch_response(
-        monkeypatch,
-        pytestconfig,
-        result_columns + license_obj.city_owned_prop_filter_cols,
-    )
+    columns = result_columns + license_obj.city_owned_prop_filter_cols
+
+    output_rows = [{c: "2020-01-01 12:00:00" for c in columns}]
+    output_rows[0]["opa_account_num"] = "1234"
+
+    maybe_monkeypatch_response(monkeypatch, pytestconfig, output_rows)
     df = license_obj.query_by_single_str_column(
         search_column="licensetype",
         search_to_match="Re",
@@ -76,3 +80,66 @@ def test_query_by_single_str_column(monkeypatch, pytestconfig):
         limit=10,
     )
     assert not df.empty
+
+
+def test_guess_property_ownership(monkeypatch, pytestconfig):
+    rtt_obj = carto_tables.RealEstateTransfers()
+    opa_account_number = "883054500"
+    maybe_monkeypatch_response(
+        monkeypatch,
+        pytestconfig,
+        [
+            {
+                "opa_account_num": opa_account_number,
+                "grantees": "COMMONWEALTH OF PENNSYLVANIA; DEPARTMENT OF TRANSPORTATION",
+                "grantors": "PHILADELPHIA ELECTRIC COMPANY",
+                "recording_date": "2016-06-12",
+                "address_low": "2301",
+                "address_low_suffix": None,
+                "address_high": None,
+                "address_low_frac": None,
+                "street_predir": None,
+                "street_name": "MARKET",
+                "street_suffix": "ST",
+                "street_address": "2301 MARKET ST",
+            },
+            {
+                "opa_account_num": opa_account_number,
+                "grantees": "",
+                "grantors": "",
+                "recording_date": "2000-01-01",
+                "address_low": "BAD ADDRESS",
+                "address_low_suffix": None,
+                "address_high": None,
+                "address_low_frac": None,
+                "street_predir": None,
+                "street_name": "MARKET",
+                "street_suffix": "ST",
+                "street_address": "2301 MARKET ST",
+            },
+        ],
+    )
+    recording_date = "2015-06-09"
+    owner_dict = rtt_obj.infer_property_ownership(
+        opa_account_number=opa_account_number, recording_date=recording_date
+    )
+    assert owner_dict["owner"] == "PHILADELPHIA ELECTRIC COMPANY"
+
+    recording_date = "2020-06-09"
+    owner_dict = rtt_obj.infer_property_ownership(
+        opa_account_number=opa_account_number, recording_date=recording_date
+    )
+    assert (
+        owner_dict["owner"]
+        == "COMMONWEALTH OF PENNSYLVANIA; DEPARTMENT OF TRANSPORTATION"
+    )
+
+    opa_account_number = "1234"
+    maybe_monkeypatch_response(
+        monkeypatch, pytestconfig, carto_rows=[],
+    )
+    recording_date = "2015-06-09"
+    owner_dict = rtt_obj.infer_property_ownership(
+        opa_account_number=opa_account_number, recording_date=recording_date
+    )
+    assert owner_dict["owner"] == None